requirements.txt
opencv-python
ultralytics
mediapipe
numpy
sounddevice
requests
python-dotenv
